LayoutTests/media/modern-media-controls/pip-button/pip-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PiPButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const pipButton = new PiPButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("pipButton.element.localName", "button");
shouldBeTrue('pipButton.element.classList.contains("icon")');
shouldBeTrue('pipButton.element.classList.contains("pip")');
shouldBeEqualToString("pipButton.iconName", `${Icons.EnterPiP}`);

const iOSPiPButton = new PiPButton({ layoutTraits: LayoutTraits.iOS });

scheduler.frameDidFire = function()
{
    if ([pipButton, iOSPiPButton].some(button => button.width == 0))
        return;

    shouldBeTrue('pipButton.element.style.webkitMaskImage.includes("macOS/pip-in@")');
    shouldBeTrue('iOSPiPButton.element.style.webkitMaskImage.includes("iOS/pip-in@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
