LayoutTests/media/modern-media-controls/pip-button/pip-button-expected.txt

Testing the PiPButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pipButton.element.localName is "button"
PASS pipButton.element.classList.contains("icon") is true
PASS pipButton.element.classList.contains("pip") is true
PASS pipButton.iconName is Icons.EnterPiP
PASS pipButton.element.style.webkitMaskImage.includes("macOS/pip-in@") became true
PASS iOSPiPButton.element.style.webkitMaskImage.includes("iOS/pip-in@") became true

PASS successfullyParsed is true

TEST COMPLETE
